+++ .gitignore
@@ -4,3 +4,4 @@
 creds.py
 /api/__pycache__/
 /api/hive13_api/__pycache__/
+/api/app/__pycache__/
