+++ api/app/__init__.py
@@ -1,0 +1,14 @@
+from flask import Flask, g, jsonify
+import json
+import random
+import hashlib
+import requests
+
+app = Flask(__name__)
+
+# import endpoint packages
+from app import test
+from app import tool
+from app import index
+from app import members
+from app import vend
