+++ api/app/tool.py
@@ -1,0 +1,5 @@
+from app import app
+
+@app.route('/api/access/tool/', methods=["GET"])
+def tool_access():
+    return {"message": "hello world"}
