+++ api/run.py
@@ -1,5 +1,5 @@
-# print("Hello, World!")
+from app import app
 
-from hive13_api import app
-
-app.run(host='0.0.0.0', port=80, debug=True)
+if __name__ == '__main__':
+    app.run(host='0.0.0.0', port=80, debug=True)
+
